feat: chat room with no auto refresh
