+++ src/app/chat/[userID]/page.tsx
@@ -1,4 +1,4 @@
-import { ChatShowcase } from "@/app/_components/create-post";
+import { ChatRoom } from "@/app/chat/components/ChatRoom";
 import { serverapi } from "@/trpc/server";
 
 export default async function Chat({ params }: { params: { userID: string } }) {
@@ -9,24 +9,14 @@ export default async function Chat({ params }: { params: { userID: string } }) {
   });
 
   return (
-    <main className="flex min-h-screen flex-col items-center justify-center bg-gradient-to-b from-[#2e026d] to-[#15162c] text-white">
-      <div className="container flex flex-col items-center justify-center gap-12 px-4 py-16 ">
-        <h1 className="text-5xl font-extrabold tracking-tight sm:text-[5rem]">
-          Turtoise <span className="text-[hsl(280,100%,70%)]">not</span> lonely!
-        </h1>
-        <div className="flex flex-col items-center gap-2">
-          <p className="text-2xl text-white">
+    <main className="flex min-h-screen flex-col items-center justify-center bg-white w-full">
+         <div className="flex flex-col items-center gap-2 fixed top-0 bg-white z-10">
+          <p className="text-2x">
             {user ? user.firstName : "Loading tRPC query..."}
           </p>
-          <p className="text-2xl text-white">{pairUser?.aka}</p>
+          <p className="text-2xl">{pairUser?.aka}</p>
         </div>
-        <ChatShowcase withUser={params.userID} />
-        <a href="/signout">
-          <button className="rounded-2xl bg-white/10 px-10 py-3 font-semibold transition hover:bg-white/20">
-            Sign Out
-          </button>
-        </a>
-      </div>
+        <ChatRoom withUser={params.userID} />
     </main>
   );
 }
