+++ src/app/chat/components/ChatMessage.tsx
@@ -7,7 +7,9 @@ import { Dayjs } from "@/utils/dayjs";
 export function ChatMessage(props: ChatMessageProps) {
   const { isMine, isShowTop, isShowBot, imageUrl } = props;
 
-  const messageBg = isMine ? "bg-red-400" : "bg-primary-400";
+  const clockTime = getClockTime(props.createdAt);
+  const messageBg = isMine ? "bg-primary-500" : "bg-neutral-100";
+  const messageColor = isMine ? "text-invert" : "text-high";
   const justifyPosition = isMine ? "justify-end" : "justify-start";
   const roundedSide = isMine
     ? "rounded-l-lg rounded-br-lg"
@@ -33,7 +35,7 @@ export function ChatMessage(props: ChatMessageProps) {
         >
           {props.senderName}
         </div>
-        <div className={cn("bg-primary-400  p-2 ", messageBg, roundedSide)}>
+        <div className={cn("bg-primary-400  p-2 ", messageBg, messageColor, roundedSide)}>
           {props.message}
         </div>
         <div className={cn("small  min-h-1 text-medium", textAlign)}>
